feat(FileSharedActions): Reorganize convertToIOFIle functions
