+++ app/src/main/java/com/infomaniak/drive/utils/PreviewUtils.kt
@@ -29,6 +29,7 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.google.gson.JsonParser
 import com.infomaniak.drive.MatomoDrive.trackFileActionEvent
 import com.infomaniak.drive.R
 import com.infomaniak.drive.data.models.File
@@ -36,10 +37,17 @@ import com.infomaniak.drive.ui.SaveExternalFilesActivity
 import com.infomaniak.drive.ui.SaveExternalFilesActivityArgs
 import com.infomaniak.drive.ui.fileList.preview.BitmapPrintDocumentAdapter
 import com.infomaniak.drive.ui.fileList.preview.PDFDocumentAdapter
-import com.infomaniak.drive.utils.PreviewPDFUtils.downloadFile
+import com.infomaniak.drive.utils.PreviewPDFUtils.PasswordProtectedException
 import com.infomaniak.drive.utils.Utils.openWith
 import com.infomaniak.drive.utils.Utils.openWithIntentExceptkDrive
+import com.infomaniak.lib.core.networking.HttpClient
+import com.infomaniak.lib.core.networking.HttpUtils
 import com.infomaniak.lib.core.utils.lightNavigationBar
+import okhttp3.Request
+import okhttp3.Response
+import java.io.BufferedInputStream
+
+private const val BUFFER_SIZE = 8192
 
 fun Activity.setupBottomSheetFileBehavior(
     bottomSheetBehavior: BottomSheetBehavior<View>,
@@ -163,22 +171,44 @@ private fun Context.printPdf(
     }
 }
 
-fun convertFileToIOFile(
+fun downloadFile(
+    externalOutputFile: IOFile,
     fileModel: File,
-    cacheFile: IOFile,
-    shouldBePdf: Boolean = false,
+    shouldBePdf: Boolean,
     onProgress: (progress: Int) -> Unit,
-): IOFile {
-    val fileNeedDownload = if (fileModel.isOnlyOfficePreview()) {
-        fileModel.isObsolete(cacheFile)
-    } else {
-        fileModel.isObsoleteOrNotIntact(cacheFile)
-    }
+) {
+    if (externalOutputFile.exists()) externalOutputFile.delete()
 
-    if (fileNeedDownload) {
-        downloadFile(cacheFile, fileModel, shouldBePdf, onProgress)
-        cacheFile.setLastModified(fileModel.getLastModifiedInMilliSecond())
-    }
+    val downloadUrl = fileModel.downloadUrl() + if (fileModel.isOnlyOfficePreview()) "?as=pdf" else ""
+    val request = Request.Builder().url(downloadUrl).headers(HttpUtils.getHeaders(contentType = null)).get().build()
+    val downloadProgressInterceptor = DownloadOfflineFileManager.downloadProgressInterceptor(onProgress = onProgress)
 
-    return cacheFile
+    val response = HttpClient.okHttpClient.newBuilder()
+        .addNetworkInterceptor(downloadProgressInterceptor)
+        .build()
+        .newCall(request)
+        .execute()
+
+    response.use {
+        if (!it.isSuccessful) {
+            val errorCode = JsonParser.parseString(it.body?.string()).asJsonObject.getAsJsonPrimitive("error").asString
+            if (errorCode == "password_protected_error") {
+                throw PasswordProtectedException()
+            } else {
+                throw Exception("Download error")
+            }
+        }
+
+        if (shouldBePdf && it.body?.contentType()?.toString() != "application/pdf") {
+            throw UnsupportedOperationException("File not supported")
+        }
+
+        createTempFile(it, externalOutputFile)
+    }
+}
+
+private fun createTempFile(response: Response, file: IOFile) {
+    BufferedInputStream(response.body?.byteStream(), BUFFER_SIZE).use { input ->
+        file.outputStream().use { output -> input.copyTo(output, BUFFER_SIZE) }
+    }
 }
